# Conflicts:
#	client/src/App.js

+++ client/src/pages/productions/components/info/ProductionScrapBtn.jsx
@@ -11,31 +11,37 @@ export default function ProductionScrapBtn() {
   const userInfo = getUser();
   const [scrapCount, setScrapCount] = useState([]);
 
+
+
   useEffect(() => {
 
-    axios({
 
-      method: "get",
-      url: `http://127.0.0.1:8000/scrap/product/${params.pid}/${userInfo.id}`
+    if (userInfo) {
 
-    }).then((result) => {
+      axios({
 
-      if (result.data.length === 1) {
+        method: "get",
+        url: `http://127.0.0.1:8000/scrap/product/${params.pid}/${userInfo.id}`
 
-        setToggle(true);
+      }).then((result) => {
 
-      }
+        if (result.data.length === 1) {
 
-      if (result.data.length === 0) {
+          setToggle(true);
 
-        setToggle(false);
+        }
 
-      }
+        if (result.data.length === 0) {
 
-    })
+          setToggle(false);
+
+        }
+
+      })
+
+    }
 
   }, []);
-
 
   useEffect(() => {
 
@@ -57,36 +63,41 @@ export default function ProductionScrapBtn() {
 
   function handleToggle() {
 
-    if (toggle === false) {
+    if (userInfo) {
 
-      axios({
+      if (toggle === false) {
 
-        method: "post",
-        url: `http://127.0.0.1:8000/scrap/product`,
-        data: { pid: params.pid, mid: userInfo.id }
+        axios({
 
-      }).then((result) => {
+          method: "post",
+          url: `http://127.0.0.1:8000/scrap/product`,
+          data: { pid: params.pid, mid: userInfo.id }
 
-        console.log("스크랩추가성공");
+        }).then((result) => {
 
-      })
+          console.log("스크랩추가성공");
+
+        })
+      }
+
+      if (toggle === true) {
+
+        axios({
+
+          method: "delete",
+          url: `http://127.0.0.1:8000/scrap/product`,
+          data: { pid: params.pid, mid: userInfo.id }
+
+        }).then((result) => {
+
+          console.log("삭제성공");
+
+        })
+
+      }
+
     }
 
-    if (toggle === true) {
-
-      axios({
-
-        method: "delete",
-        url: `http://127.0.0.1:8000/scrap/product`,
-        data: { pid: params.pid, mid: userInfo.id }
-
-      }).then((result) => {
-
-        console.log("삭제성공");
-
-      })
-
-    }
 
   }
